requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==4.2
djangorestframework==3.16.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.3.0
openai==1.23.2
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
pytest==8.2.2
pytest-django==4.8.0
openpyxl==3.1.2
